move avito specific transports to inhouse module
remove test-report dependency on report-viewer
