+++ subprojects/android-test/test-inhouse-runner/src/main/kotlin/com/avito/android/transport/ReportTransportFactory.kt
@@ -1,16 +1,16 @@
-package com.avito.android.test.report.transport
+package com.avito.android.transport
 
+import com.avito.android.test.report.transport.ExternalStorageTransport
+import com.avito.android.test.report.transport.NoOpTransport
+import com.avito.android.test.report.transport.Transport
 import com.avito.filestorage.RemoteStorage
 import com.avito.http.HttpClientProvider
 import com.avito.logger.LoggerFactory
-import com.avito.report.EntryTypeAdapterFactory
 import com.avito.report.ReportsApiFactory
 import com.avito.report.TestArtifactsProvider
 import com.avito.report.model.DeviceName
 import com.avito.report.model.ReportCoordinates
 import com.avito.time.TimeProvider
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
 
 class ReportTransportFactory(
     private val timeProvider: TimeProvider,
@@ -21,7 +21,6 @@ class ReportTransportFactory(
 ) {
 
     private val externalStorageTransport = ExternalStorageTransport(
-        gson = gson,
         timeProvider = timeProvider,
         loggerFactory = loggerFactory,
         testArtifactsProvider = testArtifactsProvider
@@ -54,11 +53,5 @@ class ReportTransportFactory(
             ReportDestination.File -> externalStorageTransport
             ReportDestination.NoOp -> NoOpTransport
         }
-    }
-
-    companion object {
-        internal val gson: Gson = GsonBuilder()
-            .registerTypeAdapterFactory(EntryTypeAdapterFactory())
-            .create()
     }
 }
