+++ subprojects/android-test/test-report/build.gradle.kts
@@ -16,10 +16,6 @@ dependencies {
     implementation(project(":common:test-report-artifacts")) {
         because("ExternalStorageTransport need to know where to store artifacts")
     }
-    implementation(project(":common:report-viewer")) {
-        // todo should be in separate, more avito-specific module
-        because("LocalRunTransport knows about AvitoReportApi")
-    }
     implementation(project(":common:logger"))
     implementation(project(":common:elastic-logger"))
     implementation(project(":common:sentry"))
