+++ subprojects/common/report-api/src/main/kotlin/com/avito/report/EntryTypeAdapterFactory.kt
